refactor(select-wallet): move unisat extension wait logic
